feat(api): expand to full PMAXTP range (16 durations × 27 probabilities)
Durations: 5min–72h (was 7, now 16).
Probabilities: 0.01%–99.9% (was 6 int, now 27 float).
Total scenarios per point: 432 (was 42).
Migration 020: alter probability column INTEGER→FLOAT, update CHECK
constraints. Requires re-running preprocess_precipitation.

+++ frontend/js/hydrograph.js
@@ -60,7 +60,7 @@
         // Use original click coordinates (inside catchment), not outlet (on boundary)
         var click = Hydrograf.app.getClickCoords();
         var duration = document.getElementById('hydro-duration').value;
-        var probability = parseInt(document.getElementById('hydro-probability').value);
+        var probability = parseFloat(document.getElementById('hydro-probability').value);
         var alpha = parseFloat(document.getElementById('hydro-alpha').value) || 2.0;
         var beta = parseFloat(document.getElementById('hydro-beta').value) || 5.0;
         var uhModel = document.getElementById('hydro-uh-model').value;
